chat_with_weather/requirements.txt
anthropic==0.42.0
python-dotenv==1.0.1
httpx==0.13.3
